@@ -11,11 +11,8 @@
         }
     ],
     "require": {
-        "php": ">= 7.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^5.0",
-        "winwin/composer-phpcs-plugin": "^0.1"
+        "php": ">= 7.2",
+        "ext-json": "*"
     },
     "autoload": {
         "psr-4": {
@@ -28,8 +25,6 @@
         }
     },
     "extra": {
-        "branch-alias": {
-            "dev-develop": "0.3-dev"
-        }
+        "source" : "git@git.cuntutu.com:kuiper/reflection.git"
     }
 }
