@@ -2,6 +2,6 @@
 
 namespace kuiper\reflection;
 
-abstract class TestCase extends \PHPUnit_Framework_TestCase
+abstract class TestCase extends \PHPUnit\Framework\TestCase
 {
 }
