@@ -2,7 +2,7 @@
 
 namespace kuiper\reflection\filter;
 
-class IntegerTypeFilterTest extends \PHPUnit_Framework_TestCase
+class IntegerTypeFilterTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @dataProvider valuesForValidate
@@ -10,7 +10,7 @@ class IntegerTypeFilterTest extends \PHPUnit_Framework_TestCase
     public function testValidate($value, $valid)
     {
         $filter = new IntegerTypeFilter();
-        $this->assertEquals($valid, $filter->validate($value));
+        $this->assertEquals($valid, $filter->isValid($value));
     }
 
     public function valuesForValidate()
