@@ -2,19 +2,23 @@
 
 namespace kuiper\reflection;
 
+use kuiper\reflection\exception\ReflectionException;
+
 interface ReflectionNamespaceInterface
 {
+    public const NAMESPACE_SEPARATOR = '\\';
+
     /**
      * Gets the namespace name.
-     *
-     * @return string
      */
-    public function getNamespace();
+    public function getNamespace(): string;
 
     /**
      * Gets all classes defined in the namespace.
      *
      * @return string[]
+     *
+     * @throws ReflectionException if file syntax error
      */
-    public function getClasses();
+    public function getClasses(): array;
 }
