@@ -11,13 +11,19 @@ class ClassType extends ReflectionType
      */
     private $className;
 
-    public function __construct($className)
+    public function __construct($className, $allowsNull = false)
     {
+        parent::__construct($allowsNull);
         $this->className = $className;
     }
 
     public function getName(): string
     {
         return $this->className;
+    }
+
+    public function isClass(): bool
+    {
+        return true;
     }
 }
